services/diseases-analyzers/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
pika==1.2.0
python-dotenv==0.19.1
